Correcciones de login y registro
